refactor(ee-acct-rt): rename module
