+++ crates/ee-acct-runtime/src/lib.rs
@@ -4,11 +4,11 @@
 
 mod exec_processing;
 mod private_input;
-mod process_update;
+mod update_processing;
 mod verification_state;
 
 pub use private_input::SharedPrivateInput;
-pub use process_update::{
+pub use update_processing::{
     apply_update_operation_unconditionally, verify_and_apply_update_operation,
 };
